Sửa login/register thành giao diện
